+++ CHANGES.md
@@ -22,6 +22,7 @@
 - `one` and `two` are now implemented on `T_container`, not just
   `T_list`.  The errors are slightly less precise, but otherwise
   nothing has changed.
+- Add `at_most_one` to complement `one` and `two`.
 
 ## Other